Adding more test and controler logic

+++ src/main/java/games/transformers/transformersapi/dao/TransformerDAO.java
@@ -39,5 +39,5 @@ public interface TransformerDAO {
      * @param id
      * @return  true if successfully deleted
      */
-    boolean deleteTransformer(long id);
+    boolean deleteTransformer(int id);
 }
