+++ src/main/java/games/transformers/transformersapi/service/TransformerService.java
@@ -42,5 +42,5 @@ public interface TransformerService {
      * @param id
      * @return true if successfully deleted
      */
-    boolean deleteTransformer(long id);
+    boolean deleteTransformer(int id);
 }
